Fix table comment not being ingested
