+++ ingestion/src/metadata/ingestion/source/database/common_db_source.py
@@ -187,11 +187,7 @@ class CommonDbSourceService(
                 f"Table description error for table [{schema_name}.{table_name}]: {exc}"
             )
         else:
-            if hasattr(table_info, "text"):
-                description = table_info["text"]
-                # DB2 connector does not return a str type
-                if isinstance(description, list):
-                    description = description[0]
+            description = table_info.get("text")
         return description
 
     def query_table_names_and_types(
