+++ ingestion/src/metadata/ingestion/source/database/db2/metadata.py
@@ -9,8 +9,11 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 """Db2 source module"""
+import traceback
+
 from ibm_db_sa.base import DB2Dialect
 from sqlalchemy.engine import reflection
+from sqlalchemy.engine.reflection import Inspector
 
 from metadata.generated.schema.entity.services.connections.database.db2Connection import (
     Db2Connection,
@@ -23,6 +26,9 @@ from metadata.generated.schema.metadataIngestion.workflow import (
 )
 from metadata.ingestion.api.source import InvalidSourceException
 from metadata.ingestion.source.database.common_db_source import CommonDbSourceService
+from metadata.utils.logger import ingestion_logger
+
+logger = ingestion_logger()
 
 
 @reflection.cache
@@ -50,3 +56,24 @@ class Db2Source(CommonDbSourceService):
                 f"Expected Db2Connection, but got {connection}"
             )
         return cls(config, metadata_config)
+
+    @staticmethod
+    def get_table_description(
+        schema_name: str, table_name: str, inspector: Inspector
+    ) -> str:
+        description = None
+        try:
+            table_info: dict = inspector.get_table_comment(table_name, schema_name)
+        # Catch any exception without breaking the ingestion
+        except Exception as exc:  # pylint: disable=broad-except
+            logger.debug(traceback.format_exc())
+            logger.warning(
+                f"Table description error for table [{schema_name}.{table_name}]: {exc}"
+            )
+        else:
+            if hasattr(table_info, "text"):
+                description = table_info["text"]
+                # DB2 connector does not return a str type
+                if isinstance(description, list):
+                    description = description[0]
+        return description
